requirements.txt
pika==1.3.2
pydantic==2.12.5
pydantic-settings==2.14.0
SQLAlchemy==2.0.49
psycopg2-binary==2.9.9
watchdog==3.0.0
boto3==1.34.0
Pillow>=10.4.0
google-genai>=1.0.0
